Do not short-circuit updateOptions() on first call from $watch
In some circumstances, the value of `scope.$eval(attrs.uiAce)` will
change between the execution of our link function and the first time
`scope.$watch(attrs.uiAce)` is triggered. Since this is
context-dependent, we cannot predict it. We could manually store
"currentOptions" only for the purpose of short-circuiting that first
call, but that would be a waste of effort, so just allow it to execute
every time.